+++ changelog.md
@@ -2,7 +2,7 @@
 
 * Add tags condition
 
-## Upcoming
+## 4.10.15
 * Invalid items are now not loaded and logged instead of breaking the whole json
 * Items `count` is now a Range object
   * So the wiki is no longer false advertisement